// FriendFinder.Api/FriendFinderSpa/src/app/models/user/signedUserDetails.ts
export class SignedUserDetails {
    id: number;
    firstName: string;
    lastName: string;
    city: string;
    country: string;
    birthDate: Date;
    email: string;
    sex: string;
    aboutMe: string;
    universityName: string;
    startDate: Date;
    finishUpDate: Date;
    hasGraduated: boolean;
    universityDesc: string;
    companyName: string;
    designation: string;
}